#23: Different Mouse behavior in Image frame

+++ formscanner-main/src/main/java/org/albertoborsetta/formscanner/commons/FormPoint.java
@@ -61,7 +61,7 @@ public class FormPoint implements Comparable<FormPoint>{
 	}
 	
 	public String toString() {
-		return "["+getX()+","+getY()+"]";
+		return "["+(int) getX()+","+(int) getY()+"]";
 	}
 	
 	public static FormPoint toPoint(String str) {
